requirements.txt
fastapi[standard]==0.135.1
python-dotenv==1.2.2
psycopg[binary]==3.3.3
sqlmodel==0.0.37